+++ src/main/java/com/purbon/kafka/topology/actions/access/builders/BuildBindingsForProducer.java
@@ -1,6 +1,6 @@
 package com.purbon.kafka.topology.actions.access.builders;
 
-import com.purbon.kafka.topology.AccessControlProvider;
+import com.purbon.kafka.topology.BindingsBuilderProvider;
 import com.purbon.kafka.topology.actions.BaseAccessControlAction;
 import com.purbon.kafka.topology.model.users.Producer;
 import java.io.IOException;
@@ -12,14 +12,14 @@ import java.util.stream.Stream;
 
 public class BuildBindingsForProducer extends BaseAccessControlAction {
 
-  private final AccessControlProvider controlProvider;
+  private final BindingsBuilderProvider builderProvider;
   private final List<Producer> producers;
   private final String fullTopicName;
 
   public BuildBindingsForProducer(
-      AccessControlProvider controlProvider, List<Producer> producers, String fullTopicName) {
+      BindingsBuilderProvider builderProvider, List<Producer> producers, String fullTopicName) {
     super();
-    this.controlProvider = controlProvider;
+    this.builderProvider = builderProvider;
     this.producers = producers;
     this.fullTopicName = fullTopicName;
   }
@@ -28,7 +28,7 @@ public class BuildBindingsForProducer extends BaseAccessControlAction {
   public void run() throws IOException {
     Stream<String> producersStream = producers.stream().map(p -> p.getPrincipal());
     bindings =
-        controlProvider.buildBindingsForProducers(
+        builderProvider.buildBindingsForProducers(
             producersStream.collect(Collectors.toList()), fullTopicName);
   }
