+++ src/main/java/com/purbon/kafka/topology/actions/access/builders/rbac/BuildPredefinedBinding.java
@@ -1,6 +1,6 @@
 package com.purbon.kafka.topology.actions.access.builders.rbac;
 
-import com.purbon.kafka.topology.AccessControlProvider;
+import com.purbon.kafka.topology.BindingsBuilderProvider;
 import com.purbon.kafka.topology.actions.BaseAccessControlAction;
 import com.purbon.kafka.topology.roles.TopologyAclBinding;
 import java.io.IOException;
@@ -10,18 +10,18 @@ import java.util.Map;
 
 public class BuildPredefinedBinding extends BaseAccessControlAction {
 
-  private final AccessControlProvider controlProvider;
+  private final BindingsBuilderProvider builderProvider;
   private final String principal;
   private final String predefinedRole;
   private final String topicPrefix;
 
   public BuildPredefinedBinding(
-      AccessControlProvider controlProvider,
+      BindingsBuilderProvider builderProvider,
       String principal,
       String predefinedRole,
       String topicPrefix) {
     super();
-    this.controlProvider = controlProvider;
+    this.builderProvider = builderProvider;
     this.principal = principal;
     this.predefinedRole = predefinedRole;
     this.topicPrefix = topicPrefix;
@@ -30,7 +30,7 @@ public class BuildPredefinedBinding extends BaseAccessControlAction {
   @Override
   public void run() throws IOException {
     TopologyAclBinding binding =
-        controlProvider.setPredefinedRole(principal, predefinedRole, topicPrefix);
+        builderProvider.setPredefinedRole(principal, predefinedRole, topicPrefix);
     bindings = Collections.singletonList(binding);
   }
